Retaining Frames and Models by using local DKV.get on a master node and then deleting local keys only
